moved ballots list to pages
