+++ web/frontend/src/pages/BallotShow.css
